Table Widget: Show null values as empty cells (#3522)
