+++ app/client/cypress/integration/Smoke_TestSuite/ClientSideTests/DisplayWidgets/Table_Property_JsonUpdate_spec.js
@@ -30,9 +30,9 @@ describe("Test Create Api and Bind to Table widget", function() {
     cy.addColumn("CustomColumn");
   });
 
-  it("Update table json data and check the column names updated", function() {
+  it("Update table json data and check the column names updated and validate empty value", function() {
     cy.SearchEntityandOpen("Table1");
-    cy.testJsontext("tabledata", JSON.stringify(this.data.TableInputUpdate));
+    cy.testJsontext("tabledata", JSON.stringify(this.data.TableInputWithNull));
     cy.wait("@updateLayout");
     cy.tableColumnDataValidation("id");
     cy.tableColumnDataValidation("email");
@@ -40,11 +40,31 @@ describe("Test Create Api and Bind to Table widget", function() {
     cy.tableColumnDataValidation("productName");
     cy.tableColumnDataValidation("orderAmount");
     cy.tableColumnDataValidation("customColumn1");
+    cy.hideColumn("id");
     cy.hideColumn("email");
     cy.hideColumn("userName");
     cy.hideColumn("productName");
-    cy.hideColumn("orderAmount");
     cy.get(".draggable-header:contains('CustomColumn')").should("be.visible");
     cy.closePropertyPane();
+    cy.readTabledataPublish("0", "0").then((tabData) => {
+      const tabValue = tabData;
+      expect(tabValue).to.be.equal("");
+    });
+  });
+
+  it("Check Selected Row(s) Resets When Table Data Changes", function() {
+    cy.isSelectRow(1);
+    cy.openPropertyPane("tablewidget");
+    cy.testJsontext("tabledata", "[]");
+    cy.wait("@updateLayout");
+    const newTableData = [...this.data.TableInput];
+    newTableData[0].userName = "";
+    cy.testJsontext("tabledata", JSON.stringify(newTableData));
+    cy.wait("@updateLayout");
+    const selectedRowsSelector = `.t--widget-tablewidget .tbody .tr.selected-row`;
+    cy.get(selectedRowsSelector).should(($p) => {
+      // should found 0 rows
+      expect($p).to.have.length(0);
+    });
   });
 });
