+++ app/client/src/widgets/TableWidget/index.tsx
@@ -20,7 +20,7 @@ import {
 } from "utils/WidgetValidation";
 import Skeleton from "components/utils/Skeleton";
 import moment from "moment";
-import { isNumber, isString, isUndefined, isEqual, xor, without } from "lodash";
+import { isNumber, isString, isNil, isEqual, xor, without } from "lodash";
 import * as Sentry from "@sentry/react";
 import { retryPromise } from "utils/AppsmithUtils";
 import withMeta from "../MetaHOC";
@@ -42,6 +42,7 @@ import {
 } from "components/designSystems/appsmith/TableComponent/Constants";
 import tablePropertyPaneConfig from "./TablePropertyPaneConfig";
 import { BatchPropertyUpdatePayload } from "actions/controlActions";
+
 const ReactTableComponent = lazy(() =>
   retryPromise(() =>
     import("components/designSystems/appsmith/TableComponent"),
@@ -334,7 +335,7 @@ class TableWidget extends BaseWidget<TableWidgetProps, WidgetState> {
                 const data =
                   isString(value) || isNumber(value)
                     ? value
-                    : isUndefined(value)
+                    : isNil(value)
                     ? ""
                     : JSON.stringify(value);
                 tableRow[accessor] = data;
